most warnings squashed

+++ miscutils.c
@@ -28,7 +28,7 @@ void *mmalloc(size_t n)
 
   p = (void *)malloc(n);
   if (!p) {
-    fprintf(stderr, "Not enough memory to allocate %d bytes\n", n);
+    fprintf(stderr, "Not enough memory to allocate %d bytes\n", (int)n);
     abort();
   }
   return p;
